pyproject.toml
[project]
name = "dynamic-soft-augmentation"
version = "0.1.0"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "optuna>=3.4.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scipy>=1.12.0",
]
[build-system]
requires = ["setuptools", "wheel"]
